Shared schemas moved to libs/types/src/property/ (schema.ts + types.ts + index.ts), exported as the property namespace from libs/types/src/index.ts.
