+++ libs/types/src/index.ts
@@ -11,6 +11,7 @@ export * from './linear-types'
 export * as keys from './keys'
 export * as leasing from './leasing'
 export * as economy from './economy'
+export * as property from './property'
 export * as schemas from './schemas'
 export { paginatedResponseSchema } from './schemas/pagination'
 export * as fileStorageSchemas from './schemas/file-storage'
