+++ libs/types/src/property/index.ts
@@ -1,0 +1,2 @@
+export * from './schema'
+export * from './types'
