+++ services/property/package.json
@@ -36,6 +36,7 @@
   "dependencies": {
     "@koa/cors": "^5.0.0",
     "@koa/router": "^12.0.1",
+    "@onecore/types": "workspace:*",
     "@onecore/utilities": "workspace:*",
     "@prisma/client": "^5.21.1",
     "axios": "^1.9.0",
