+++ services/property/src/types/curves.ts
@@ -46,53 +46,10 @@ export const EcoGuardDataSubNodeSchema = z.object({
 
 export const EcoGuardDataResponseSchema = z.array(EcoGuardDataSubNodeSchema)
 
-// ---- Public schemas (request + response shape) ----
-
-export const ApartmentTemperaturesIntervalSchema = z.enum(['H', 'D'])
-
-export const ApartmentTemperaturesQuerySchema = z
-  .object({
-    from: z.coerce.number().int().positive().optional(),
-    to: z.coerce.number().int().positive().optional(),
-    interval: ApartmentTemperaturesIntervalSchema.optional(),
-  })
-  .refine((q) => q.from === undefined || q.to === undefined || q.to > q.from, {
-    message: '`to` must be greater than `from`',
-    path: ['to'],
-  })
-
-export const ApartmentTemperaturePointSchema = z.object({
-  time: z.number(),
-  avg: z.number().nullable(),
-  min: z.number().nullable(),
-  max: z.number().nullable(),
-})
-
-export const ApartmentTemperatureSeriesSchema = z.object({
-  subNodeId: z.number(),
-  subNodeName: z.string(),
-  points: z.array(ApartmentTemperaturePointSchema),
-})
-
-export const ApartmentTemperaturesResponseSchema = z.object({
-  objectNumber: z.string(),
-  nodeId: z.number(),
-  from: z.number(),
-  to: z.number(),
-  interval: ApartmentTemperaturesIntervalSchema,
-  unit: z.string(),
-  series: z.array(ApartmentTemperatureSeriesSchema),
-})
+// Public request/response schemas live in `@onecore/types` (property module)
+// as the single source of truth shared with core. Only the EcoGuard upstream
+// shapes stay here.
 
 export type EcoGuardTokenResponse = z.infer<typeof EcoGuardTokenResponseSchema>
 export type EcoGuardApartmentNode = z.infer<typeof EcoGuardApartmentNodeSchema>
 export type EcoGuardDataResponse = z.infer<typeof EcoGuardDataResponseSchema>
-export type ApartmentTemperaturesQuery = z.infer<
-  typeof ApartmentTemperaturesQuerySchema
->
-export type ApartmentTemperaturesInterval = z.infer<
-  typeof ApartmentTemperaturesIntervalSchema
->
-export type ApartmentTemperaturesResponse = z.infer<
-  typeof ApartmentTemperaturesResponseSchema
->
